ai-module/requirements.txt
flask==2.3.3
flask-cors==4.0.0
nltk==3.8.1
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0